pyproject.toml
[project]
name = "mcp-client-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.68.0",
    "fastapi>=0.117.1",
    "mcp>=1.14.1",
    "python-dotenv>=1.1.1",
]
